+++ php_mf2parse.h
@@ -11,36 +11,11 @@
   +----------------------------------------------------------------------+
 */
 
-#ifndef PHP_MF2_H
-#define PHP_MF2_H
+#ifndef PHP_MF2PARSE_H
+#define PHP_MF2PARSE_H
 
-#if HAVE_MF2
+extern zend_class_entry *mf2parse_ce;
 
-#ifdef ZTS
-#include "TSRM.h"
-#endif
+PHP_MINIT_FUNCTION( mf2parse );
 
-#define PHP_MF2_VERSION "0.1.0"
-
-extern zend_module_entry mf2_module_entry;
-#define mf2_module_ptr &mf2_module_entry
-
-#include "mf2_str_defs.h"
-
-typedef struct _mf2_str_globals {
-#define X(str) zend_string *str_ ## str;
-	MF2_STR_DEFS
-#undef X
-} mf2_str_globals;
-
-mf2_str_globals str_globals_mf2;
-
-#define MF2_STR(str) str_globals_mf2.str
-
-#else
-#define mf2_module_ptr NULL
-#endif /* HAVE_MF2 */
-
-#define phpext_mf2_ptr mf2_module_ptr
-
-#endif /* PHP_MF2 */
+#endif /* PHP_MF2PARSE_H */