+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ## Unreleased
 
 ### Added
+- TEST: MF2Parse class exists
 - TEST: phpinfo() includes mf2 status
 - TEST: mf2 extension is loaded
 - Minimal PHP extension config